moved load_inputvariablesfile into bifacial_radiance.load
